move the `TableRow` class into its own file
